made background color reactive. Also tidied up code in Components folder and separated styles from component
